pyproject.toml
[project]
name = "building-production-ai-systems"
version = "0.1.0"
description = "RAG API with FastAPI + in-memory Qdrant, using OpenAI embeddings"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.6",
    "openai==1.51.0",
    "httpx==0.27.0",
    "qdrant-client==1.11.3",
    "langchain-core==0.3.67",
    "langchain-community==0.3.26",
    "langchain-openai",
    "langchain-qdrant==0.2.0",
    "langchain-text-splitters",
    "langgraph==0.5.0",
    "pymupdf>=1.26.1",
    "python-dotenv==1.0.1",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.1.1",
    "pandas>=2.3.0",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "langchain-ollama",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["app*"]
exclude = ["tests*"]
